# eyewear_node_typescript/src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Eyewear API"
  },
  "basePath": "/api/v1",
  "parameters": {
    "Authrization": {
      "in": "header",
      "name": "authorization",
      "type": "string",
      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Indhc2VlbUB5b3BtYWlsLmNvbSIsInVzZXJJZCI6IjYxNmE4MmQyNzQwZTk5MzY4OTNmZTdiNCIsImlhdCI6MTYzNzY2NzkxMSwiZXhwIjoxNjQwMDg3MTExfQ.sOiVb2KjS_5mfUJSPKi9Iymx4mr8LJb3UElGuZ3999Q",
      "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6Indhc2VlbUB5b3BtYWlsLmNvbSIsInVzZXJJZCI6IjYxNmE4MmQyNzQwZTk5MzY4OTNmZTdiNCIsImlhdCI6MTYzNzY2NzkxMSwiZXhwIjoxNjQwMDg3MTExfQ.sOiVb2KjS_5mfUJSPKi9Iymx4mr8LJb3UElGuZ3999Q"
    },
    "Timezone": {
      "in": "header",
      "name": "timezone",
      "type": "string",
      "example": "Asia/Calcutta",
      "default": "Asia/Calcutta",
      "required": true
    },
    "AcceptLanguage": {
      "in": "header",
      "name": "Accept-Language",
      "type": "string",
      "example": "en/hi",
      "default": "hi",
      "required": true
    }
  },
...
  "paths": {
...
    "/register": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "User Registration",
        "description": "User register api",
        "tags": ["User auth apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "required": true,
                  "type": "string",
                  "example": "Don Joe"
                },
                "email": {
                  "required": true,
                  "type": "string",
                  "example": "donjoe@gmail.com"
                },
                "password": {
                  "required": true,
                  "type": "string",
                  "example": "Test@123"
                },
                "login_type": {
                  "required": true,
                  "type": "number",
                  "example": 0
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/userRegister"
            }
          }
        }
      }
    },
    "/sendotp": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "resend OTP",
        "description": "resned OTP API",
        "tags": ["User auth apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "required": true,
                  "type": "string",
                  "example": "example@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/resendOtp"
            }
          }
        }
      }
    },
    "/verifyotp": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "verify OTP",
        "description": "verify OTP API",
        "tags": ["User auth apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "required": true,
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "otp": {
                  "required": true,
                  "type": "number",
                  "example": 1234
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/verifyOtp"
            }
          }
        }
      }
    },
    "/login": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "user login",
        "description": "Login Api",
        "tags": ["User auth apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "required": true,
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "password": {
                  "required": true,
                  "type": "string",
                  "example": "Test@123"
                },
                "login_type": {
                  "required": true,
                  "type": "number",
                  "example": 0
                },
                "device_token": {
                  "required": true,
                  "type": "string",
                  "example": "evooy2mXFUCsSk60mSohrg:APA91bEsfv3OTo8bJw80hyDTOwxSK-uWEHwFKOg1hRIQs0PWgW2msso3p4XylM9M-bIn9Hu-dQ8dArQ1Sa2TaRJ0uFkMWJcGRYz_97WbVoZXiDfjcQNNUbfmDYH9cK8RIqwN66xu3uNv"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/loginSave"
            }
          }
        }
      }
    },
    "/sociallogin": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "user login",
        "description": "Login Api",
        "tags": ["User auth apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "exampleone"
                },
                "email": {
                  "type": "string",
                  "example": "example@gmail.com"
                },
                "social_id": {
                  "required": true,
                  "type": "string",
                  "example": "117988134179193593975"
                },
                "login_type": {
                  "required": true,
                  "type": "number",
                  "example": 0
                },
                "device_token": {
                  "required": true,
                  "type": "string",
                  "example": "evooy2mXFUCsSk60mSohrg:APA91bEsfv3OTo8bJw80hyDTOwxSK-uWEHwFKOg1hRIQs0PWgW2msso3p4XylM9M-bIn9Hu-dQ8dArQ1Sa2TaRJ0uFkMWJcGRYz_97WbVoZXiDfjcQNNUbfmDYH9cK8RIqwN66xu3uNv"
                }
              }
            }
          }
        ]
      },
      "responses": {
        "200": {
          "description": "successfull operation",
          "schema": {
            "$ref": "#/definitions/socialloginSave"
          }
        }
      }
    },
    "/forgotpassword": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "forgot password",
        "description": "forgot password API",
        "tags": ["User auth apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "required": true,
                  "type": "string",
                  "example": "example@gmail.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/forgotPassword"
            }
          }
        }
      }
    },
    "/refreshtoken": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "refresh token",
        "description": "refesh token ApI",
        "tags": ["User auth apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "required": true,
                  "type": "string",
                  "example": "6655bf4a002885850f5392cf"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/refreshToken"
            }
          }
        }
      }
    },
...
    "/carts": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "get": {
        "summary": "get user cart api",
        "description": "get user cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "type": "string",
            "required": true,
            "example": "616a82d2740e9936893fe7b4",
            "default": "616a82d2740e9936893fe7b4"
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartGet"
            }
          }
        }
      }
    },
    "/carts/": {
      "parameters": [
        {
          "$ref": "#/parameters/Authrization"
        },
        {
          "$ref": "#/parameters/Timezone"
        },
        {
          "$ref": "#/parameters/AcceptLanguage"
        }
      ],
      "post": {
        "summary": "add cart api",
        "description": "add cart API",
        "tags": ["Cart apis"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "user_id": {
                  "type": "string",
                  "example": "616a82d2740e9936893fe7b4"
                },
                "product_id": {
                  "type": "string",
                  "example": "616325ececd4223283d4f123"
                },
                "quantity": {
                  "required": true,
                  "type": "number",
                  "example": 1
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successfull operation",
            "schema": {
              "$ref": "#/definitions/cartAdd"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "userRegister": {
      "allOf": [
        {
          "$ref": "#/definitions/userRegister"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "successfullly registered"
            }
          }
        }
      ]
    },
    "resendOtp": {
      "allOf": [
        {
          "$ref": "#/definitions/resendOtp"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "otp resended successfullly"
            }
          }
        }
      ]
    },
    "verifyOtp": {
      "allOf": [
        {
          "$ref": "#/definitions/verifyOtp"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "otp verified successfullly"
            }
          }
        }
      ]
    },
    "loginSave": {
      "allOf": [
        {
          "$ref": "#/definitions/loginSave"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "logged in successfullly"
            }
          }
        }
      ]
    },
    "socialloginSave": {
      "allOf": [
        {
          "$ref": "#/definitions/socialloginSave"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "logged in successfullly"
            }
          }
        }
      ]
    },
    "forgotPassword": {
      "allOf": [
        {
          "$ref": "#/definitions/forgotPassword"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "password resend on email"
            }
          }
        }
      ]
    },
    "refreshToken": {
      "allOf": [
        {
          "$ref": "#/definitions/refreshToken"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "data retrived successfullly"
            }
          }
        }
      ]
    },
...
    "cartGet": {
      "allOf": [
        {
          "$ref": "#/definitions/cartGet"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "cart retrived successfullly!"
            }
          }
        }
      ]
    },
    "cartAdd": {
      "allOf": [
        {
          "$ref": "#/definitions/cartAdd"
        },
        {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "add to cart successfullly!"
            }
          }
        }
      ]
    },
...
  }
}